# openapi_backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimVision-API",
    "description": "API del proyecto ClaimVision, en constante actualizacion y mejoria",
    "version": "1.12.0"
  },
  "paths": {
...
    "/api/v1/aseguradora/crud/talleres": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Taller",
        "operationId": "crear_taller_api_v1_aseguradora_crud_talleres_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallerCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Listar Talleres",
        "operationId": "listar_talleres_api_v1_aseguradora_crud_talleres_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TallerResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/talleres/{id}": {
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Obtener Taller",
        "operationId": "obtener_taller_api_v1_aseguradora_crud_talleres__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Actualizar Taller",
        "operationId": "actualizar_taller_api_v1_aseguradora_crud_talleres__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallerUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Eliminar Taller",
        "operationId": "eliminar_taller_api_v1_aseguradora_crud_talleres__id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/talleres/{id}/operadores": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Operador Taller",
        "operationId": "crear_operador_taller_api_v1_aseguradora_crud_talleres__id__operadores_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperadorTallerRequestDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/talleres": {
      "get": {
        "tags": [
          "v1 · Admin"
        ],
        "summary": "Listar Talleres Admin",
        "description": "Listar todos los talleres del sistema (sin filtro por aseguradora).",
        "operationId": "listar_talleres_admin_api_v1_admin_talleres_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TallerAdminResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/talleres/{taller_id}": {
      "get": {
        "tags": [
          "v1 · Admin"
        ],
        "summary": "Obtener Taller Admin",
        "description": "Obtener detalle de un taller con aseguradoras vinculadas.",
        "operationId": "obtener_taller_admin_api_v1_admin_talleres__taller_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerAdminResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperadorTallerRequestDTO": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "puesto": {
            "type": "string",
            "title": "Puesto",
            "default": "Operador"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "email",
          "password"
        ],
        "title": "OperadorTallerRequestDTO"
      },
...
      "Page_TallerAdminResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TallerAdminResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[TallerAdminResponseDTO]"
      },
      "Page_TallerResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TallerResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[TallerResponseDTO]"
      },
...
      "TallerAdminResponseDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nombre_comercial": {
            "type": "string",
            "title": "Nombre Comercial"
          },
          "rfc": {
            "type": "string",
            "title": "Rfc"
          },
          "direccion_tecnica": {
            "type": "string",
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "type": "string",
            "title": "Telefono Contacto"
          },
          "aseguradoras_vinculadas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aseguradoras Vinculadas",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre_comercial",
          "rfc",
          "direccion_tecnica",
          "telefono_contacto",
          "created_at",
          "updated_at"
        ],
        "title": "TallerAdminResponseDTO"
      },
      "TallerCreateDTO": {
        "properties": {
          "nombre_comercial": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombre Comercial"
          },
          "rfc": {
            "type": "string",
            "maxLength": 13,
            "minLength": 12,
            "pattern": "^[A-ZÑ&]{3,4}\\d{6}[A-Za-z\\d]{3}$",
            "title": "Rfc"
          },
          "direccion_tecnica": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Telefono Contacto"
          }
        },
        "type": "object",
        "required": [
          "nombre_comercial",
          "rfc",
          "direccion_tecnica",
          "telefono_contacto"
        ],
        "title": "TallerCreateDTO"
      },
...
      "TallerResponseDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nombre_comercial": {
            "type": "string",
            "title": "Nombre Comercial"
          },
          "rfc": {
            "type": "string",
            "title": "Rfc"
          },
          "direccion_tecnica": {
            "type": "string",
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "type": "string",
            "title": "Telefono Contacto"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre_comercial",
          "rfc",
          "direccion_tecnica",
          "telefono_contacto",
          "version",
          "created_at",
          "updated_at",
          "deleted_at"
        ],
        "title": "TallerResponseDTO"
      },
      "TallerUpdateDTO": {
        "properties": {
          "nombre_comercial": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Comercial"
          },
          "direccion_tecnica": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono Contacto"
          }
        },
        "type": "object",
        "title": "TallerUpdateDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}